+++ velux-scenes.js
@@ -1,9 +1,9 @@
 module.exports = function (RED) {
   'use strict'
   var util = require('util')
-  var debug = require('debug')('node-red-contrib-velux:velux-nodes')
+  var debug = require('debug')('node-red-contrib-velux:velux-scenes')
 
-  function veluxNodes(config) {
+  function veluxScenes(config) {
     RED.nodes.createNode(this, config)
     var node = this
 
@@ -11,7 +11,7 @@ module.exports = function (RED) {
     if (node.veluxDatasource) {
       node.hasTopic = (config.topic||'').length > 0
       debug('config:',config)
-      node.veluxDatasource.subscribeNodes(node)
+      //node.veluxDatasource.subscribeScenes(node)
       
       node.getID = function(){
         return config.index
@@ -21,37 +21,41 @@ module.exports = function (RED) {
         return node.veluxDatasource.nodeGetName(node.getID)
       }
 
+      node.getVelocity = function(){
+        return node.veluxDatasource.getVelocityTagByName(config.velocity)
+      }
+      
+      var timer = null
+      node.runTimer = function(seconds){
+        clearTimeout(timer)
+        var s = seconds
+        timer = setInterval(()=>{
+          //n.status({fill:fillSystem,shape:shapeSystem,text:textSystem})
+          node.status({text:""+--seconds+"sec"})
+          if (seconds<1) {
+            clearTimeout(timer)
+          }
+        },1000)
+      }
+      
       node.publish = function(data){
         debug('node.send:',(!(!data)))
         if (data) {
+          if (data.apiText === "GW_COMMAND_REMAINING_TIME_NTF") {
+            node.runTimer(data.seconds)
+          }
+          if (data.apiText === "GW_COMMAND_RUN_STATUS_NTF") {
+            clearTimeout(timer)
+            node.status({text:data.runStatusText})
+          }
           var msg={}
           if (node.hasTopic) {
             msg.topic = config.topic
           } else {
-            msg.topic = data.nodeName
+            msg.topic = node.name
           }
 
-          if (config.nodevalue == 'CURRENTPOSITION') {
-            msg.payload = data.currentPosition.value
-            msg.valueType = data.currentPosition.valueType
-          } else if (config.nodevalue == 'TARGET') {
-            msg.payload = data.target.value
-            msg.valueType = data.target.valueType
-          } else if (config.nodevalue == 'FP1') {
-            msg.payload = data.fp1CurrentPosition.value
-            msg.valueType = data.fp1CurrentPosition.valueType
-          } else if (config.nodevalue == 'FP2') {
-            msg.payload = data.fp2CurrentPosition.value
-            msg.valueType = data.fp2CurrentPosition.valueType
-          } else if (config.nodevalue == 'FP3') {
-            msg.payload = data.fp3CurrentPosition.value
-            msg.valueType = data.fp3CurrentPosition.valueType
-          } else if (config.nodevalue == 'FP4') {
-            msg.payload = data.fp4CurrentPosition.value
-            msg.valueType = data.fp4CurrentPosition.valueType
-          } else {
-            msg.payload = data
-          } 
+          msg.payload = data
 
           node.send(msg)
         }
@@ -60,79 +64,63 @@ module.exports = function (RED) {
       node.on("input", function(msg) {
         debug('input:','msg',msg)
         
-        if (typeof msg.topic === 'string') {
+        if (typeof msg.topic === 'string' && msg.topic !== '') {
+          debug('input:','msg.topic === string')
           if (node.hasTopic && msg.topic == (config.topic||'')){
-            node.veluxDatasource.nodeSendValue(node.getID(),msg.payload)
+            node.veluxDatasource.runScene(node,node.getID(),node.getVelocity())
+            .then((data)=>{node.publish(data)})
+            .catch((err)=>{node.error(util.format('Velux %s', err))})
             return
           }
           var topicvals = msg.topic.split(":")
           if ((topicvals[0]||'').trim()== 'velux') {
-            if (((topicvals[1]||'').trim()=='read')||((topicvals[1]||'').trim()=='load')) {
-              var load = (topicvals[1].trim()=='load')
-              if ((topicvals[2]||'').trim()=='id') {
-                var id = parseInt(topicvals[3]||'')
-                debug ('input:',id)
-                if (id==-1) {
-                  node.veluxDatasource.nodePublishAll(node, load)
-                } else {
-                  node.veluxDatasource.nodePublish(node, id, load)
-                }
-              } else if ((topicvals[2]||'').trim()=='name') {
-                var len = topicvals[0].length + topicvals[1].length + topicvals[2].length + 3
-                var name = msg.topic.substring(len).trim()
-                var id = node.veluxDatasource.getIdByName(name)
-                debug ('input:',name,id)
-                if (id === null) return
-                node.veluxDatasource.nodePublish(node, id,  load)
-              } else {
-                if (node.getID()==-1) {
-                  node.veluxDatasource.nodePublishAll(node, load)
-                } else {
-                  node.veluxDatasource.nodePublish(node, node.getID(), load)
-                }
-              }
-            } else if ((topicvals[1]||'').trim()=='write') {
-              var val = {value: parseFloat(msg.payload), valueType: 'RELATIVE'}
+            if ((topicvals[1]||'').trim()=='execute') {
               var id = node.getID()
-              var name
-              for (var i=2;i<topicvals.length;i++) {
-                if ((topicvals[i]||'').trim()=='valuetype') {
+              var velocity = node.getVelocity()
+              for (var i=1;i<topicvals.length;i++) {
+                if ((topicvals[i]||'').trim()=='id') {
                   i++
-                  val.valueType = (topicvals[i]||'').trim()
-                } else if ((topicvals[i]||'').trim()=='rawvalue') {
-                  val = {rawValue: parseFloat(msg.payload)}
-                } else if ((topicvals[i]||'').trim()=='id') {
-                  i++
-                  id = parseInt(topicvals[i]||'')
+                  var text = (topicvals[i]||'')
+                  debug('input:','id',text)
+                  id = parseInt(text)
                 } else if ((topicvals[i]||'').trim()=='name') {
                   i++
-                  name = (topicvals[i]||'')
+                  var text = (topicvals[i]||'')
+                  debug('input:','name',text)
+                  id = node.veluxDatasource.getSceneByName(text)
+                } else if ((topicvals[i]||'').trim()=='velocityid') {
+                  i++
+                  var text = (topicvals[i]||'')
+                  debug('input:','velocityid',text)
+                  velocity = parseInt(text)
+                } else if ((topicvals[i]||'').trim()=='velocity') {
+                  i++
+                  var text = (topicvals[i]||'')
+                  debug('input:','velocity',text)
+                  velocity = node.veluxDatasource.getVelocityTagByName(text)
                 } 
               }
-              debug('input:','name',name)
-              if (name) {
-                id = node.veluxDatasource.getIdByName(name)
-                if (id === null) return
-              }
-              debug('input:','name',name)
-              node.veluxDatasource.nodeSendValue(id,val)
+              node.veluxDatasource.runScene(node,id,velocity)
+              .then((data)=>{node.publish(data)})
+              .catch((err)=>{node.error(util.format('Velux %s', err))})
             }
             return
           }
           if (msg.topic) return
-        } 
-        if (!isNaN(parseFloat(msg.payload))) {
-          node.veluxDatasource.nodeSendValue(node.getID(),parseFloat(msg.payload))
+        } else {
+          node.veluxDatasource.runScene(node,node.getID(),node.getVelocity())
+          .then((data)=>{node.publish(data)})
+          .catch((err)=>{node.error(util.format('Velux %s', err))})
           return
         }
       })
       
       node.on('close', function (done) {
         debug('close:')
-        node.veluxDatasource.unsubscribeNodes(node)
+        //node.veluxDatasource.unsubscribeScenes(node)
         done()
       })
     }
   }
-  RED.nodes.registerType('Velux Nodes', veluxNodes)
+  RED.nodes.registerType('Velux Scenes', veluxScenes)
 }