+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 - velux nodes: added remaining time
 - velux api: added
+- velux scenes: added scenes
 
 
 **Fixes**
 
-- 
+- velux api: corrected topic
 
 
 ### 0.0.1: Initial Release